% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convK.R
\name{plot.convKern}
\alias{plot.convKern}
\title{\code{plot} method for class \code{convKern}}
\usage{
\method{plot}{convKern}(x, ...)
}
\arguments{
\item{x}{An object of \code{convKern} class}

\item{...}{further arguments to pass to \code{plot} method}
}
\description{
\code{plot} method for class \code{convKern}
}

