% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convK.R
\name{print.convKern}
\alias{print.convKern}
\title{\code{print} method for class \code{convKern}}
\usage{
print.convKern(x, ...)
}
\arguments{
\item{x}{An object of \code{convKern} class}

\item{...}{further arguments to pass to \code{print} method}
}
\description{
\code{print} method for class \code{convKern}
}

